test(cli): move to e2e specs under test and simplify fixtures
