+++ packages/libretto-cli/test/basic.spec.ts
@@ -1,30 +1,35 @@
 import { existsSync } from "node:fs";
 import { mkdir, readFile, writeFile } from "node:fs/promises";
+import { join } from "node:path";
 import { describe, expect } from "vitest";
-import { test } from "./test-fixtures";
+import { test } from "./fixtures";
 
 describe("basic CLI subprocess behavior", () => {
   test("prints usage for --help", async ({ librettoCli }) => {
     const result = await librettoCli("--help");
     expect(result.exitCode).toBe(0);
-    expect(result.stdout).toContain("Usage: libretto-cli <command> [--session <name>]");
+    expect(result.stdout).toContain(
+      "Usage: libretto-cli <command> [--session <name>]",
+    );
     expect(result.stderr).toBe("");
   });
 
   test("bootstraps .libretto state on --help without creating legacy dirs", async ({
     librettoCli,
-    workspacePath,
+    workspaceDir,
   }) => {
     const result = await librettoCli("--help");
     expect(result.exitCode).toBe(0);
 
-    expect(existsSync(workspacePath(".libretto"))).toBe(true);
-    expect(existsSync(workspacePath(".libretto", ".gitignore"))).toBe(true);
-    expect(existsSync(workspacePath(".libretto", "sessions"))).toBe(true);
-    expect(existsSync(workspacePath(".libretto", "profiles"))).toBe(true);
+    expect(existsSync(join(workspaceDir, ".libretto"))).toBe(true);
+    expect(existsSync(join(workspaceDir, ".libretto", ".gitignore"))).toBe(
+      true,
+    );
+    expect(existsSync(join(workspaceDir, ".libretto", "sessions"))).toBe(true);
+    expect(existsSync(join(workspaceDir, ".libretto", "profiles"))).toBe(true);
 
-    expect(existsSync(workspacePath(".libretto-cli"))).toBe(false);
-    expect(existsSync(workspacePath("tmp", "libretto-cli"))).toBe(false);
+    expect(existsSync(join(workspaceDir, ".libretto-cli"))).toBe(false);
+    expect(existsSync(join(workspaceDir, "tmp", "libretto-cli"))).toBe(false);
   });
 
   test("prints usage for help command", async ({ librettoCli }) => {
@@ -34,11 +39,15 @@ describe("basic CLI subprocess behavior", () => {
     expect(result.stderr).toBe("");
   });
 
-  test("fails unknown command with non-zero exit code", async ({ librettoCli }) => {
+  test("fails unknown command with non-zero exit code", async ({
+    librettoCli,
+  }) => {
     const result = await librettoCli("nope-command");
     expect(result.exitCode).toBe(1);
     expect(result.stderr).toContain("Unknown command: nope-command");
-    expect(result.stdout).toContain("Usage: libretto-cli <command> [--session <name>]");
+    expect(result.stdout).toContain(
+      "Usage: libretto-cli <command> [--session <name>]",
+    );
   });
 
   test("fails open with missing url usage error", async ({ librettoCli }) => {
@@ -75,15 +84,14 @@ describe("basic CLI subprocess behavior", () => {
     const result = await librettoCli("run ./integration.ts main");
     expect(result.exitCode).toBe(1);
     expect(result.stderr).toContain(
-      "Session \"default\" is read-only. Only a human can authorize interactive mode.",
+      'Session "default" is read-only. Only a human can authorize full-access mode.',
     );
   });
 
-  test("allows run guard when session is permissioned interactive", async ({
+  test("allows run guard when session is permissioned full-access", async ({
     librettoCli,
-    seedSessionPermission,
   }) => {
-    await seedSessionPermission("default", "full-access");
+    await librettoCli("session-mode full-access --session default");
     const result = await librettoCli("run ./integration.ts main");
     expect(result.exitCode).toBe(1);
     expect(result.stderr).not.toContain("is read-only");
@@ -92,18 +100,16 @@ describe("basic CLI subprocess behavior", () => {
 
   test("fails run when export is not a Libretto workflow instance", async ({
     librettoCli,
-    seedSessionPermission,
-    workspacePath,
+    writeWorkflow,
   }) => {
-    await seedSessionPermission("default", "full-access");
-    await writeFile(
-      workspacePath("integration.ts"),
+    await librettoCli("session-mode full-access --session default");
+    await writeWorkflow(
+      "integration.ts",
       `
 export async function main() {
   return "ok";
 }
 `,
-      "utf8",
     );
 
     const result = await librettoCli("run ./integration.ts main");
@@ -113,12 +119,12 @@ export async function main() {
 
   test("accepts branded Libretto workflow contract across module boundaries", async ({
     librettoCli,
-    seedSessionPermission,
-    workspacePath,
+    workspaceDir,
+    writeWorkflow,
   }) => {
-    await seedSessionPermission("default", "full-access");
-    await writeFile(
-      workspacePath("integration.ts"),
+    await librettoCli("session-mode full-access --session default");
+    await writeWorkflow(
+      "integration.ts",
       `
 const brand = Symbol.for("libretto.workflow");
 
@@ -130,11 +136,13 @@ export const main = {
   },
 };
 `,
-      "utf8",
     );
 
     const result = await librettoCli("run ./integration.ts main", {
-      PLAYWRIGHT_BROWSERS_PATH: workspacePath("missing-playwright-browsers"),
+      PLAYWRIGHT_BROWSERS_PATH: join(
+        workspaceDir,
+        "missing-playwright-browsers",
+      ),
     });
     expect(result.exitCode).toBe(1);
     expect(result.stderr).not.toContain("must be a Libretto workflow instance");
@@ -142,19 +150,12 @@ export const main = {
 
   test("fails run when local auth profile is declared but missing", async ({
     librettoCli,
-    seedSessionPermission,
-    workspacePath,
+    writeWorkflow,
   }) => {
-    const librettoEntryUrl = new URL(
-      "../../libretto/src/workflow/workflow.ts",
-      import.meta.url,
-    ).href;
-    await seedSessionPermission("default", "full-access");
-    await writeFile(
-      workspacePath("integration.ts"),
+    await librettoCli("session-mode full-access --session default");
+    await writeWorkflow(
+      "integration.ts",
       `
-import { workflow } from "${librettoEntryUrl}";
-
 export const main = workflow(
   { authProfile: { type: "local", domain: "app.example.com" } },
   async () => {
@@ -162,7 +163,6 @@ export const main = workflow(
   },
 );
 `,
-      "utf8",
     );
 
     const result = await librettoCli("run ./integration.ts main");
@@ -170,101 +170,92 @@ export const main = workflow(
     expect(result.stderr).toContain(
       'Local auth profile not found for domain "app.example.com".',
     );
-    expect(result.stderr).toContain(
-      "Expected profile file:",
-    );
-    expect(result.stderr).toContain(
-      ".libretto/profiles/app.example.com.json",
-    );
+    expect(result.stderr).toContain("Expected profile file:");
+    expect(result.stderr).toContain(".libretto/profiles/app.example.com.json");
     expect(result.stderr).toContain(
       "libretto-cli open https://app.example.com --headed --session default",
     );
-    expect(result.stderr).toContain("libretto-cli save app.example.com --session default");
+    expect(result.stderr).toContain(
+      "libretto-cli save app.example.com --session default",
+    );
   });
 
   test("does not require local auth profile when auth metadata is absent", async ({
     librettoCli,
-    seedSessionPermission,
-    workspacePath,
+    workspaceDir,
+    writeWorkflow,
   }) => {
-    const librettoEntryUrl = new URL(
-      "../../libretto/src/workflow/workflow.ts",
-      import.meta.url,
-    ).href;
-    await seedSessionPermission("default", "full-access");
-    await writeFile(
-      workspacePath("integration.ts"),
+    await librettoCli("session-mode full-access --session default");
+    await writeWorkflow(
+      "integration.ts",
       `
-import { workflow } from "${librettoEntryUrl}";
-
 export const main = workflow({}, async () => "ok");
 `,
-      "utf8",
     );
 
     const result = await librettoCli("run ./integration.ts main", {
-      PLAYWRIGHT_BROWSERS_PATH: workspacePath("missing-playwright-browsers"),
+      PLAYWRIGHT_BROWSERS_PATH: join(
+        workspaceDir,
+        "missing-playwright-browsers",
+      ),
     });
     expect(result.exitCode).toBe(1);
-    expect(result.stderr).not.toContain("Local auth profile not found for domain");
+    expect(result.stderr).not.toContain(
+      "Local auth profile not found for domain",
+    );
   });
 
   test("proceeds when declared local auth profile file exists", async ({
     librettoCli,
-    seedSessionPermission,
-    workspacePath,
+    workspaceDir,
+    writeWorkflow,
   }) => {
-    const librettoEntryUrl = new URL(
-      "../../libretto/src/workflow/workflow.ts",
-      import.meta.url,
-    ).href;
-    await seedSessionPermission("default", "full-access");
-    await writeFile(
-      workspacePath("integration.ts"),
+    await librettoCli("session-mode full-access --session default");
+    await writeWorkflow(
+      "integration.ts",
       `
-import { workflow } from "${librettoEntryUrl}";
-
 export const main = workflow(
   { authProfile: { type: "local", domain: "app.example.com" } },
   async () => "ok",
 );
 `,
-      "utf8",
     );
-    await mkdir(workspacePath(".libretto", "profiles"), { recursive: true });
+    await mkdir(join(workspaceDir, ".libretto", "profiles"), {
+      recursive: true,
+    });
     await writeFile(
-      workspacePath(".libretto", "profiles", "app.example.com.json"),
+      join(workspaceDir, ".libretto", "profiles", "app.example.com.json"),
       JSON.stringify({ cookies: [], origins: [] }),
       "utf8",
     );
 
     const result = await librettoCli("run ./integration.ts main", {
-      PLAYWRIGHT_BROWSERS_PATH: workspacePath("missing-playwright-browsers"),
+      PLAYWRIGHT_BROWSERS_PATH: join(
+        workspaceDir,
+        "missing-playwright-browsers",
+      ),
     });
     expect(result.exitCode).toBe(1);
-    expect(result.stderr).not.toContain("Local auth profile not found for domain");
+    expect(result.stderr).not.toContain(
+      "Local auth profile not found for domain",
+    );
   });
 
   test("returns paused status when workflow hits debugPause", async ({
     librettoCli,
-    librettoRuntimePath,
-    seedSessionPermission,
-    writeWorkflowScript,
+    writeWorkflow,
   }) => {
-    await seedSessionPermission("default", "interactive");
-    const integrationFilePath = await writeWorkflowScript(
+    await librettoCli("session-mode full-access --session default");
+    const integrationFilePath = await writeWorkflow(
       "integration-pause.mjs",
       `
-import { workflow, debugPause } from "${
-  librettoRuntimePath
-}";
-
 export const main = workflow({}, async (ctx) => {
   console.log("WORKFLOW_BEFORE_PAUSE");
   await debugPause(ctx.page, { enabled: ctx.debug, sessionName: ctx.session });
   console.log("WORKFLOW_AFTER_PAUSE");
 });
 `,
+      ["workflow", "debugPause"],
     );
 
     const result = await librettoCli(
@@ -279,18 +270,12 @@ export const main = workflow({}, async (ctx) => {
 
   test("completes workflow run when no pause is triggered", async ({
     librettoCli,
-    librettoRuntimePath,
-    seedSessionPermission,
-    writeWorkflowScript,
+    writeWorkflow,
   }) => {
-    await seedSessionPermission("default", "interactive");
-    const integrationFilePath = await writeWorkflowScript(
+    await librettoCli("session-mode full-access --session default");
+    const integrationFilePath = await writeWorkflow(
       "integration-complete.mjs",
       `
-import { workflow } from "${
-  librettoRuntimePath
-}";
-
 export const main = workflow({}, async () => {
   console.log("WORKFLOW_COMPLETES");
 });
@@ -309,7 +294,9 @@ export const main = workflow({}, async () => {
   test("fails open when deprecated --allow-actions flag is passed", async ({
     librettoCli,
   }) => {
-    const result = await librettoCli("open https://example.com --allow-actions");
+    const result = await librettoCli(
+      "open https://example.com --allow-actions",
+    );
     expect(result.exitCode).toBe(1);
     expect(result.stderr).toContain(
       "--allow-actions is not supported for open.",
@@ -319,26 +306,27 @@ export const main = workflow({}, async () => {
   test("fails run when deprecated --allow-actions flag is passed", async ({
     librettoCli,
   }) => {
-    const result = await librettoCli("run ./integration.ts main --allow-actions");
+    const result = await librettoCli(
+      "run ./integration.ts main --allow-actions",
+    );
     expect(result.exitCode).toBe(1);
-    expect(result.stderr).toContain("--allow-actions is not supported for run.");
+    expect(result.stderr).toContain(
+      "--allow-actions is not supported for run.",
+    );
   });
 
-  test("session-mode interactive writes session permission", async ({
+  test("session-mode full-access writes session permission", async ({
     librettoCli,
-    workspacePath,
+    workspaceDir,
   }) => {
     const result = await librettoCli(
-      "session-mode interactive --session consented",
+      "session-mode full-access --session consented",
     );
     expect(result.exitCode).toBe(0);
-    expect(result.stdout).toContain("Session \"consented\" is now interactive.");
+    expect(result.stdout).toContain('Session "consented" is now full-access.');
 
     const raw = JSON.parse(
-      await readFile(
-        workspacePath(".libretto", "config.json"),
-        "utf8",
-      ),
+      await readFile(join(workspaceDir, ".libretto", "config.json"), "utf8"),
     ) as {
       permissions?: {
         sessions?: Record<string, string>;
@@ -347,20 +335,19 @@ export const main = workflow({}, async () => {
     expect(raw.permissions?.sessions?.consented).toBe("full-access");
   });
 
-  test("session-mode read-only removes interactive permission", async ({
+  test("session-mode read-only removes full-access permission", async ({
     librettoCli,
-    workspacePath,
+    workspaceDir,
   }) => {
-    await librettoCli("session-mode interactive --session toggled");
-    const result = await librettoCli("session-mode read-only --session toggled");
+    await librettoCli("session-mode full-access --session toggled");
+    const result = await librettoCli(
+      "session-mode read-only --session toggled",
+    );
     expect(result.exitCode).toBe(0);
-    expect(result.stdout).toContain("Session \"toggled\" is now read-only.");
+    expect(result.stdout).toContain('Session "toggled" is now read-only.');
 
     const raw = JSON.parse(
-      await readFile(
-        workspacePath(".libretto", "config.json"),
-        "utf8",
-      ),
+      await readFile(join(workspaceDir, ".libretto", "config.json"), "utf8"),
     ) as {
       permissions?: {
         sessions?: Record<string, string>;
@@ -373,11 +360,13 @@ export const main = workflow({}, async () => {
     const result = await librettoCli("session-mode maybe --session default");
     expect(result.exitCode).toBe(1);
     expect(result.stderr).toContain(
-      "Usage: libretto-cli session-mode <read-only|interactive> [--session <name>]",
+      "Usage: libretto-cli session-mode <read-only|full-access> [--session <name>]",
     );
   });
 
-  test("fails save with missing target usage error", async ({ librettoCli }) => {
+  test("fails save with missing target usage error", async ({
+    librettoCli,
+  }) => {
     const result = await librettoCli("save");
     expect(result.exitCode).toBe(1);
     expect(result.stderr).toContain(
