+++ packages/libretto-cli/test/stateful.spec.ts
@@ -1,7 +1,49 @@
 import { existsSync } from "node:fs";
 import { mkdir, readFile, writeFile } from "node:fs/promises";
+import { join } from "node:path";
 import { describe, expect } from "vitest";
-import { test } from "./test-fixtures";
+import { test } from "./fixtures";
+
+async function writeSessionState(
+  workspaceDir: string,
+  state: {
+    session: string;
+    runId?: string;
+    port?: number;
+    pid?: number;
+    startedAt?: string;
+    mode?: "read-only" | "full-access";
+  },
+): Promise<string> {
+  const dir = join(workspaceDir, ".libretto", "sessions", state.session);
+  const path = join(workspaceDir, ".libretto", "sessions", state.session, "state.json");
+  await mkdir(dir, { recursive: true });
+  const payload: Record<string, unknown> = {
+    version: 1,
+    session: state.session,
+    runId: state.runId ?? "run-seeded",
+    port: state.port ?? 9222,
+    pid: state.pid ?? 12345,
+    startedAt: state.startedAt ?? "2026-01-01T00:00:00.000Z",
+  };
+  if (state.mode) payload.mode = state.mode;
+  await writeFile(path, JSON.stringify(payload, null, 2), "utf8");
+  return path;
+}
+
+async function writeJsonl(
+  workspaceDir: string,
+  session: string,
+  fileName: "network.jsonl" | "actions.jsonl",
+  entries: Record<string, unknown>[],
+): Promise<string> {
+  const sessionDir = join(workspaceDir, ".libretto", "sessions", session);
+  const filePath = join(workspaceDir, ".libretto", "sessions", session, fileName);
+  await mkdir(sessionDir, { recursive: true });
+  const body = entries.map((entry) => JSON.stringify(entry)).join("\n");
+  await writeFile(filePath, body ? `${body}\n` : "", "utf8");
+  return filePath;
+}
 
 describe("state-driven CLI subprocess behavior", () => {
   test("shows missing AI config", async ({ librettoCli }) => {
@@ -12,13 +54,13 @@ describe("state-driven CLI subprocess behavior", () => {
 
   test("configures, shows, and clears AI config", async ({
     librettoCli,
-    workspacePath,
+    workspaceDir,
   }) => {
     const configure = await librettoCli("ai configure codex");
     expect(configure.exitCode).toBe(0);
     expect(configure.stdout).toContain("AI config saved.");
 
-    const configPath = workspacePath(".libretto", "config.json");
+    const configPath = join(workspaceDir, ".libretto", "config.json");
     expect(existsSync(configPath)).toBe(true);
     const rawConfig = JSON.parse(await readFile(configPath, "utf8")) as {
       ai?: {
@@ -45,13 +87,13 @@ describe("state-driven CLI subprocess behavior", () => {
 
   test("configures gemini AI preset", async ({
     librettoCli,
-    workspacePath,
+    workspaceDir,
   }) => {
     const configure = await librettoCli("ai configure gemini");
     expect(configure.exitCode).toBe(0);
     expect(configure.stdout).toContain("AI config saved.");
 
-    const configPath = workspacePath(".libretto", "config.json");
+    const configPath = join(workspaceDir, ".libretto", "config.json");
     const rawConfig = JSON.parse(await readFile(configPath, "utf8")) as {
       ai?: {
         preset?: string;
@@ -67,16 +109,14 @@ describe("state-driven CLI subprocess behavior", () => {
   });
 
   test("reads and clears network logs from seeded run data", async ({
-    seedSessionState,
-    seedNetworkLog,
     librettoCli,
-    workspacePath,
+    workspaceDir,
   }) => {
-    await seedSessionState({
+    await writeSessionState(workspaceDir, {
       session: "net-session",
       runId: "run-net",
     });
-    const logPath = await seedNetworkLog("net-session", [
+    const logPath = await writeJsonl(workspaceDir, "net-session", "network.jsonl", [
       {
         ts: "2026-01-01T00:00:00.000Z",
         method: "GET",
@@ -112,21 +152,19 @@ describe("state-driven CLI subprocess behavior", () => {
     const cleared = await readFile(logPath, "utf8");
     expect(cleared).toBe("");
     expect(
-      existsSync(workspacePath(".libretto", "sessions", "net-session", "network.jsonl")),
+      existsSync(join(workspaceDir, ".libretto", "sessions", "net-session", "network.jsonl")),
     ).toBe(true);
   });
 
   test("reads and clears action logs from seeded run data", async ({
-    seedSessionState,
-    seedActionLog,
     librettoCli,
-    workspacePath,
+    workspaceDir,
   }) => {
-    await seedSessionState({
+    await writeSessionState(workspaceDir, {
       session: "actions-session",
       runId: "run-actions",
     });
-    const logPath = await seedActionLog("actions-session", [
+    const logPath = await writeJsonl(workspaceDir, "actions-session", "actions.jsonl", [
       {
         ts: "2026-01-01T00:00:00.000Z",
         action: "click",
@@ -161,16 +199,16 @@ describe("state-driven CLI subprocess behavior", () => {
     expect(cleared).toBe("");
     expect(
       existsSync(
-        workspacePath(".libretto", "sessions", "actions-session", "actions.jsonl"),
+        join(workspaceDir, ".libretto", "sessions", "actions-session", "actions.jsonl"),
       ),
     ).toBe(true);
   });
 
   test("blocks exec in read-only open sessions", async ({
-    seedSessionState,
+    workspaceDir,
     librettoCli,
   }) => {
-    await seedSessionState({
+    await writeSessionState(workspaceDir, {
       session: "readonly-session",
       mode: "read-only",
     });
@@ -180,15 +218,15 @@ describe("state-driven CLI subprocess behavior", () => {
     );
     expect(result.exitCode).toBe(1);
     expect(result.stderr).toContain(
-      "Session \"readonly-session\" is read-only. Only a human can authorize interactive mode.",
+      "Session \"readonly-session\" is read-only. Only a human can authorize full-access mode.",
     );
   });
 
   test("rejects exec when session mode is not specified", async ({
-    seedSessionState,
+    workspaceDir,
     librettoCli,
   }) => {
-    await seedSessionState({
+    await writeSessionState(workspaceDir, {
       session: "legacy-session",
     });
 
@@ -197,38 +235,53 @@ describe("state-driven CLI subprocess behavior", () => {
     );
     expect(result.exitCode).toBe(1);
     expect(result.stderr).toContain(
-      "Session \"legacy-session\" is read-only. Only a human can authorize interactive mode.",
+      "Session \"legacy-session\" is read-only. Only a human can authorize full-access mode.",
     );
     expect(result.stderr).toContain(
-      "libretto-cli session-mode interactive --session legacy-session",
+      "libretto-cli session-mode full-access --session legacy-session",
     );
   });
 
-  test("rejects exec when session mode is missing even if permissioned interactive", async ({
-    seedSessionState,
-    seedSessionPermission,
+  test("rejects exec when session mode is missing even if permissioned full-access", async ({
+    workspaceDir,
     librettoCli,
   }) => {
-    await seedSessionState({
+    await writeSessionState(workspaceDir, {
       session: "permissioned-session",
       port: 65534,
     });
-    await seedSessionPermission("permissioned-session", "full-access");
+    await mkdir(join(workspaceDir, ".libretto"), { recursive: true });
+    await writeFile(
+      join(workspaceDir, ".libretto", "config.json"),
+      JSON.stringify(
+        {
+          version: 1,
+          permissions: {
+            sessions: {
+              "permissioned-session": "full-access",
+            },
+          },
+        },
+        null,
+        2,
+      ),
+      "utf8",
+    );
 
     const result = await librettoCli(
       "exec \"return await page.title()\" --session permissioned-session",
     );
     expect(result.exitCode).toBe(1);
     expect(result.stderr).toContain(
-      "Session \"permissioned-session\" is read-only. Only a human can authorize interactive mode.",
+      "Session \"permissioned-session\" is read-only. Only a human can authorize full-access mode.",
     );
   });
 
   test("does not apply read-only guard when session allows actions", async ({
-    seedSessionState,
+    workspaceDir,
     librettoCli,
   }) => {
-    await seedSessionState({
+    await writeSessionState(workspaceDir, {
       session: "interactive-session",
       port: 65534,
       mode: "full-access",
@@ -245,10 +298,10 @@ describe("state-driven CLI subprocess behavior", () => {
   });
 
   test("rejects exec after session is switched back to read-only", async ({
-    seedSessionState,
+    workspaceDir,
     librettoCli,
   }) => {
-    await seedSessionState({
+    await writeSessionState(workspaceDir, {
       session: "flip-session",
       mode: "full-access",
     });
@@ -266,22 +319,20 @@ describe("state-driven CLI subprocess behavior", () => {
     );
     expect(result.exitCode).toBe(1);
     expect(result.stderr).toContain(
-      "Session \"flip-session\" is read-only. Only a human can authorize interactive mode.",
+      "Session \"flip-session\" is read-only. Only a human can authorize full-access mode.",
     );
-    expect(result.stderr).toContain(
-      "libretto-cli session-mode interactive --session flip-session",
-    );
+    expect(result.stderr).toContain("libretto-cli session-mode full-access --session flip-session");
   });
 
   test("rejects session state files with unsupported version", async ({
     librettoCli,
-    workspacePath,
+    workspaceDir,
   }) => {
     const session = "invalid-version-session";
-    const sessionDir = workspacePath(".libretto", "sessions", session);
+    const sessionDir = join(workspaceDir, ".libretto", "sessions", session);
     await mkdir(sessionDir, { recursive: true });
     await writeFile(
-      workspacePath(".libretto", "sessions", session, "state.json"),
+      join(workspaceDir, ".libretto", "sessions", session, "state.json"),
       JSON.stringify(
         {
           version: 2,
