+++ packages/libretto-cli/tsup.config.ts
@@ -1,7 +1,7 @@
 import { defineConfig } from "tsup";
 
 export default defineConfig({
-	entry: ["src/**/*.ts", "!src/**/*.test.ts", "!src/test-fixtures.ts"],
+	entry: ["src/**/*.ts", "!src/**/*.test.ts", "!src/**/*.spec.ts"],
 	format: ["esm"],
 	dts: false,
 	clean: true,
